refactor migrations: add renaming migration
