+++ jobs/mongodb_migration/src/mongodb_migration/migration.py
@@ -3,6 +3,17 @@
 
 import datetime
 from abc import ABC, abstractmethod
+from typing import Any
+
+from libcommon.constants import (
+    CACHE_COLLECTION_RESPONSES,
+    CACHE_MONGOENGINE_ALIAS,
+    METRICS_COLLECTION_CACHE_TOTAL_METRIC,
+    METRICS_COLLECTION_JOB_TOTAL_METRIC,
+    METRICS_MONGOENGINE_ALIAS,
+    QUEUE_COLLECTION_JOBS,
+    QUEUE_MONGOENGINE_ALIAS,
+)
 
 
 class IrreversibleMigrationError(Exception):
@@ -31,3 +42,32 @@ class Migration(ABC):
     @abstractmethod
     def down(self) -> None:
         raise IrreversibleMigrationError()
+
+
+class QueueMigration(Migration, ABC):
+    MONGOENGINE_ALIAS: str = QUEUE_MONGOENGINE_ALIAS
+    COLLECTION_JOBS: str = QUEUE_COLLECTION_JOBS
+
+    def __init__(self, job_type: str, *args: Any, **kwargs: Any):
+        self.job_type = job_type
+        super().__init__(*args, **kwargs)
+
+
+class CacheMigration(Migration, ABC):
+    MONGOENGINE_ALIAS: str = CACHE_MONGOENGINE_ALIAS
+    COLLECTION_RESPONSES: str = CACHE_COLLECTION_RESPONSES
+
+    def __init__(self, cache_kind: str, *args: Any, **kwargs: Any):
+        self.cache_kind = cache_kind
+        super().__init__(*args, **kwargs)
+
+
+class MetricsMigration(Migration, ABC):
+    MONGOENGINE_ALIAS: str = METRICS_MONGOENGINE_ALIAS
+    COLLECTION_JOB_TOTAL_METRIC: str = METRICS_COLLECTION_JOB_TOTAL_METRIC
+    COLLECTION_CACHE_TOTAL_METRIC: str = METRICS_COLLECTION_CACHE_TOTAL_METRIC
+
+    def __init__(self, job_type: str, cache_kind: str, *args: Any, **kwargs: Any):
+        self.job_type = job_type
+        self.cache_kind = cache_kind
+        super().__init__(*args, **kwargs)
